refactor: factor out types and dataclasses
